backend/requirements.txt
fastapi
uvicorn
python-socketio
massive
mcp
